-chg: ????????? ???? SciTE-Ru ????????? ?? ?????? 2.12 ????????? ???????? ??? ?????????? ? MinGW: ../src/Editor.cxx: In member function 'void Editor::PaintSelMargin(Surface*, PRectangle&)': ../src/Editor.cxx:1788: warning: suggest explicit braces to avoid ambiguous 'else' ../src/KeyWords.cxx: In member function 'bool WordList::InMultiWordsList(const char*, char, bool&, bool&, const char*&)': ../src/KeyWords.cxx:319: warning: suggest explicit braces to avoid ambiguous 'else' ../src/LexEuphoria.cxx: In function 'void ColouriseEuphoriaDoc(unsigned int, int, int, WordList**, Accessor&)': ../src/LexEuphoria.cxx:141: warning: suggest parentheses around '&&' within '||' ../src/LexLua.cxx: In function 'void ColouriseLuaDoc(unsigned int, int, int, WordList**, Accessor&)': ../src/LexLua.cxx:355: warning: suggest explicit braces to avoid ambiguous 'else' ../src/LexSQL.cxx: In function 'void FoldSQLDoc(unsigned int, int, int, WordList**, Accessor&)': ../src/LexSQL.cxx:303: warning: suggest parentheses around '&&' within '||' ../src/LuaExtension.cxx:328: warning: unused parameter 'L'
????? ???????????

+++ src/scintilla/win32/PlatWin.cxx
@@ -2121,10 +2121,7 @@ double ElapsedTime::Duration(bool reset) {
 		LARGE_INTEGER lBegin;
 		lBegin.HighPart = bigBit;
 		lBegin.LowPart = littleBit;
-//!-start-[no_wornings]
-//		double elapsed = lEnd.QuadPart - lBegin.QuadPart;
-		double elapsed = static_cast<double>(lEnd.QuadPart - lBegin.QuadPart);
-//!-end-[no_wornings]
+		double elapsed = lEnd.QuadPart - lBegin.QuadPart;
 		result = elapsed / static_cast<double>(frequency.QuadPart);
 	} else {
 		endBigBit = clock();
